本次提交完成了两个需求：   1：聚合session，一天session在各个时间跨度区间，步长跨度区间的分布率,并插入数据库   2：抽取的session总数记为100个,按每个小时占得一天总session的比例，随机抽取session,并插入数据库

+++ src/main/scala/empcl/dao/impl/TaskDaoImpl.scala
@@ -2,9 +2,9 @@ package empcl.dao.impl
 
 import java.sql.{PreparedStatement, ResultSet}
 
-import empcl.cases.Task
 import empcl.dao.ITaskDao
 import empcl.helper.JdbcPoolHelper
+import empcl.spark.Task
 
 /**
   * @author : empcl
@@ -31,7 +31,7 @@ class TaskDaoImpl extends ITaskDao {
 
     val parseResultSet = (rs: ResultSet) => {
       if (rs.next()) {
-        val taskId = rs.getLong(1)
+        val taskId = rs.getInt(1)
         val taskName = rs.getString(2)
         val createTime = rs.getString(3)
         val startTime = rs.getString(4)
